Add files via upload
// Team Happy Coding 
// Ramish Masood, 209512, MTS-B
// Tazah Gul, 229713, MTS-A
// Muhammad Umer, 220671, MTS-B

VEHICLE VERIFICATION ANDROID APPLICATION

The following android app was created by us as the course project of Data Structures & Object Oriented Programming (DS&OOP). 
Our app works in the following manner:
1. Splash Screen for 3 seconds
2. Main Page consists of 3 buttons which are camera, data, about us.
3. In data, only admin registers the user. After adding member, user is allowed to to insert data, delete data, update data and view data appears.
The camera opens and picture of number plate is took and the app scans the number on number plate. This number is matched with the database and the car is allowed to enter if result is positive. Otherwise, the vehicle is denied entrance. Extension includes a blacklist in which number of stolen cars is added and when that number is detected, civil security institutes are notified. 
This app can be installed in academic institutes, hospitals, corporations and wherever identification is required to enter.